Now using the GA to identify the parameter values of a real, experimental system.
While this has not been fully tested, a few bugs when using the GA found in the pysysid package have been fixed, which makes the possible convergence of the algorithm with real world data more and more convincing.
A full test still needs to be completed to determine is the algorithm is still able to perform with noisy data, which will not exactly match the expected output of the given process model.